Fix a complie error in generated go files
Fixes #67.

protoc-gen-go started to keep unused imports of .proto files.
So we need to map google/api/annotation.proto to a valid golang import path.

Another possible solution was to map the proto file to some harmless package,
e.g. "fmt".  But it would be safer to map it to a valid path.