README: Suggest go get -u by default.
Suggesting go get -u is a safer default. It results in a more predictable action, the user will end up with the latest version of grpc-gateway and all its dependencies. Otherwise, the versions users end up with will be arbitrary, which may lead to issues. If users prefer not to update dependencies, they can always go get without -u flag.

As an example of the issue this would help resolve, see #103.