Fix #119 - CloseNotify race with ServeHTTP
The generated code for the mux handlers can panic if the call to
ServeHTTP() finishes before the goroutine finishes. This is because
CloseNotify() can only be called before ServeHTTP() finishes. I've
fixed this by calling CloseNotify() before starting the goroutine. The
returned channel can safely be selected any time after that.