Generate Swagger description for service methods using proto comments.
While this is a first step in resolving gengo/grpc-gateway#128, this
needs to be cleaned up, and the same approach needs to be used for
messages, message fields, et al.

echo_service.proto has been annotated with extra comments in order to
demo the new descriptions.

Only the Swagger example has been regenerated, as my local generator
does not output all the expected fields in proto struct tags.