Swagger: summary, description, custom JSON, package docs.
This patch will extract summary separately from description in
Swagger objects that happen to support both summary and
description. It will do so by splitting the relevant proto
comment based on paragraphs, and using the first paragraph
as the summary.

This patch will also allow updating the Swagger schema's
otherwise hard-to-map fields by allowing custom JSON to be
placed within the proto comment. The syntax leaves something to
be desired, but it works. Use of JSON is, on the other hand,
rather universal.

This patch will also allow describing the whole API by attaching
the comments to the 'package' stanza inside the proto. In this
case, summary will be applied to info.title, and description
will be applied to info.description. Other properties have to be
set through JSON.

Schema has been slightly updated to allow for fields such as
termsOfService, contact, license, externalDocs, etc.

If a method is not documented, stub summary for Swagger
Operation is no longer generated.

Documentation for enum values is now multiline with dashes
before values.