runtime: Use protobuf enum name instead of (reflect.Type).String
This adds pass-through of the *proto.Properties for fields so that the
fields being populated can be addressed by their enum name when the
type is an enum (i.e., (*proto.Properties).Enum is set).

As a result of this, the tests now pass, using the protobuf enum name
as registered (i.e., fully qualified package name) instead of the Go
package name derived by calling (reflect.Type).String. This should have
no real impact on existing use, but it may be justified to re-add
a fallback case for that reflect.Type name for hand-written protobuf
types. Even then, however, it seems unlikely that it should be expected
for enums-by-name to work when registering the enum map under a name
other than the one it's referred to in protobuf.