runtime: Add tests for oneof query values
This adds tests, which fail, to read values for oneof fields from the
query string. The assumption for the change right now is that the query
string should only contain one use of a oneof group's fields -- e.g.,
in the test, only one of oneof_string_value or oneof_bool_value, or
neither, may be specified. Passing both is an error.

This also adds a wanterr field to the test spec for parameters, simply
to detect that case.