Add custom options to allow more control of swagger/openapi output (#145)
Creates additional options that can be applied to RPCs and messages that provide additional information in the OpenAPIv2 spec. These new options allow for the inclusion of 

Swagger object information (maps to a .proto file):
 * `info`
 * `host`
 * `base_path`
 * `schemes`
 * `consumes`
 * `produces`
 * `external_docs`

Service information:
 * `tags`
 * `summary`
 * `description`
 * `external_docs`
 * `operation_id`
 * `consumes`
 * `produces`
 * `schemes`

Operation information (maps to a method):
 * `tags`
 * `summary`
 * `description`
 * `external_docs`
 * `operation_id`
 * `consumes`
 * `produces`
 * `schemes`

Schema information (maps to a message):
 * `json_schema`
 * `descriptor`
 * `read_only`
 * `external_docs`
 * `example`

Additionally:
* Change from 'Swagger: ' to 'OpenAPI: ' prefix.

It should also be relatively simple to add any field that is missing to the new OpenAPI proto objects.