Add unannotated example for e2e testing
To make this work we have to adjust the makefile to pass
the additional grpc api configuration file. We also have
to provide a basic server implementation and use the
generated client from the tests.

The e2e test is basically just the echo example but with an
additional inclusion of the google Duration WKT to ensure
we properly generate the imports for those even if the annotation
WKT is not used.