Add support for basic gRPC API Configuration YAML files (#521)
* Add support for basic gRPC API Configuration YAML files

Separate gRPC API Configuration files as described by google at
https://cloud.google.com/endpoints/docs/grpc/grpc-service-config
can be used to separate the description of the basic API surface
in .proto files from their mapping to HTTP REST. In grpc-gateway
this allows exposing gRPC services without having to add any
annotations to their .proto files.

The format of gRPC API Configuration YAML files is defined by the
proto message schema google.api.service which can be found as
service.proto in the googleapis repository. This patch exposes
the Http part of the service configuration to the gateway
generator.

To enable this all HttpRules contained in the Http part of the
description are loaded into the generator registry as a map from
service method to HttpRules called externalHttpRules. During
loading of the proto files we can now use these externalHttpRules
to look for additional rules to apply to a service method in addition
to the ones from the annotations found in the proto file. As the
types are the same no fundamentally new codepaths are required in
either of the generators.

google.api.service is quite a complex protobuf message with lots
of dependencies on other google api. To not have to pull in all
other dependent types this patch constructs a reduced service
message type with only the Http field. This relies on protobufs
backwards compatibility guarantees which allows us to remove all
other fields. Also we do not need all protobuf features on this
custom message.

To load the actual YAML file as a protobuf message this patch uses
a new dependency on github.com/ghodss/yaml to convert the YAML into
JSON which we can then unmarshal into protobuf using jsonpb.

* Add a "Generating for unannotated protos" section to README.md

Create a new documentation page describing how to use
grpc-gateway without annotations by the use of gRPC
API Configuration files. This new section explains the
use of gRPC API Configuration YAML files to expose
unannotated .proto files in contrast to the annotation
approach.

Adjusted other parts of the documentation to hint/link
at that capability.

* Add unannotated example for e2e testing

The e2e test is basically just the echo example but with an
additional inclusion of the google Duration WKT to ensure
we properly generate the imports for those even if the annotation
WKT is not used.

To make this work we have to adjust the makefile to pass
the additional grpc api configuration file. We also have
to provide a basic server implementation and use the
generated client from the tests.