Generate import mappings.
Instead of generating a list of import paths based on the difnames of the
imported protos, generate maps to the package local names. This is needed to
handle the case of packages containing a proto with the same name.

An example here is the annotation.proto in this module.

Fixes #633