Add test case and proposed fix for path component with trailing colon (and string) (#708)
* If a pattern has no verb, match with a verb arg appended to last component

Parsing out a "verb" from the input path is a convenience for
parsing. Whether the colon and string are _actually_ a verb depends on
the matching HTTP rule. This change gives a verb-less pattern a chance
to match a path by assuming the colon+string suffix are a part of the
final component.

Fixes #224

Signed-off-by: James Hamlin <james@goforward.com>