Use newer Bazel repo rules (#798)
This uses the new go_repository and http_archive rules. Using the former
avoids a warning:

DEPRECATED: the rules_go version of go_repository is deprecated. Migrate
to the new version in @bazel_gazelle//def.bzl instead.

The latter will avoid breakage when the native http_archive rule is
disabled by default in a future version of Bazel.