write Description Definition into Parameter.Description
Some tools render the field paths.parameters.description of the swagger JSON.
this field is currently not set. there is also no possibility to explicitly set
this through options. As this is directly a reference to the definition of the
message type and the NAME of the message type (=> swagger definition) is also
provided, there is no harm in filling the description field as well.

Example:

"paths": {
  "/v1/user": {
    "post": {
      "summary": "...",
      "operationId": "...",
      "responses": {
        "200": {
          "description": "A successful response.",
          "schema": {
            "$ref": "#/definitions/protobufBoolValue"
          }
        }
      },
      "parameters": [
        {
          "name": "body",
          "description": "Creating a user",
          "in": "body",
          "required": true,
          "schema": {
            "$ref": "#/definitions/CreateUserRequest"
          }
        }
      ],
...

The field "parameters.description" is missing without this patch. We just
post-process the swagger struct, peek into the correct definition and get the
description.

The implementation is by far not perfect as the $ref string is parsed for the
definition name. I couldn't find a better location though, there was no place
with access to the requires structures.