Run proto gen swagger with protos from external repository
Currently the protoc_gen_swagger doesn't work with protos from external repositories.
The bazel rule because the swagger.json file is not stored at the expected place.
Example if the rule for the proto @remote//proto:action_proto is passed to the protoc_gen_swagger,
the action.swagger.json file is written to proto/action.swagger.json and
is expected to be at the same place as the proto external/remote/proto/action.swagger.json.