Fix hand-crafted protobuf message
errorBody is an improperly hand-crafted protobuf Message.
In particular, field number 1 is used twice, causing this to
message to crash in the v2 re-implementation of Go protobufs.

This PR modifies the field numbers to be unique.
Since Error is a special field not in status.proto,
we choose a large field number to avoid conflicting with
any new field that may be added to the real Status message.
We also fix the field numbers to truly match up with Status.