Error on nested namespacing collisions for message
Nested messages can currently have name collisions with un-nested
messages, this commit adds an error if a collision occurs, though
hopefully this should be a reasonably rare occurrence.

In the case of a message Bar with a nested field Foo colliding with a
message BarFoo the error message will look like:

```
<input>:1:1:protoc-gen-swagger: message schema name collision: testBarFoo would be defined twice due to message nesting
```

It is possible there may be other causes for collision but I am
reasonably sure this is the only way it can happen without a failure in
protoc first.