Create tracing.md
This is a first basic draft for adding tracing. This is for using OpenCensus and aws x-ray so maybe it should be added to another section/file. There are no screenshots in this one, I could use the ones from https://github.com/johanbrandhorst/grpc-gateway-boilerplate/tree/add-tracing however that is from another repo. It would not be to hard to add an actual example of this in /examples/internal. Please let me know what to change/remove.