feat: make default error handlers public
Pre-v2 it was possible to define "error handling middleware" that
modified errors before delegating to the default error handler
(runtime.DefaultHTTPProtoErrorHandler).

In our case, we were using this capability to hook in a custom request
error logger.

By exposing the default error handlers, this remains possible on the v2
branch.