Deprecate example field (#1637)
* Update contribution guide to use 1.15 generator

* Deprecate example field in favor of example_string

The example field using google.protobuf.Any
was pretty silly and just happened to work because
of how we were using it. It's confusing to users and breaks some tooling,
so it is best deprecated.

This is an unfortunate break from the "each field maps to the openapiv2 schema"
precedent we've maintained so far, but there is
no other good way that I can think of to safely deprecate
this.