Migrate generation, linting to buf (#1971)
* Migrate to buf

Migrate to buf for generation, linting and breaking
change detection. Removes much of the complicated
Makefile logic and the need for protoc
in favour of buf generate.

* Regenerate files with buf

* Remove local googleapis copy

This is no longer needed since we can depend
on googleapis from the BSR

* Update README to be in line with local practices

* Update the docs for BSR availability

* Exclude bazel directory

The generated, hidden bazel directory will
mess up buf's generation if it exists, so we ignore it.