Handle custom status code mappings for routing errors (#2101)
* Handle custom status code mappings for routing errors

Not all gRPC error codes and http status codes can have a one to one
mapping. Http can have both a path and a method component, while gRPC
will just have a function that is specific to the combination of method
and path when translating between the two. That means that for gRPC a
missing function simply means "unimplemented", while for the equivalent
REST API it is common to distinguish an unhandled method request on a
known resource as "Method Not Allowed" with the response code 405.

This may not be the only place where expectations of response codes
needed may not align with what is provided by the current mapping
chain http.Status* -> gRPC code.Codes -> http.Status*, where some
information may be lost in the double conversion.

As the error handler accepts an 'error' interface, it should be possible
to provide a custom error structure that can contain the http Status
code to wrap the original error to allow for the expected response code
to be provided without impacting the current function signatures.

* update with goimports

* Restore DefaultRoutingErrorHandler and tests, and update doc strings

Remove the customization from the DefaultRoutingErrorHandler as this is
what should be done by consumers, consequently remove the modifications
to tests affected.

Update doc strings as requested.

* Rename StatusHTTP to HTTPStatus

* Update names and docstrings

* Update customizing your gateway docs

* rework example

* remove newline